Moved hooks folder
Moved hooks folder to src.
Moved firebase folder into hooks folder.
Removed `auth.signout()` since it was for testing purpose.
